Update to v106r118 release.
byuu says:

  - added watchdog timer emulation,
  - improved interrupt handling,
  - implemented the GPIO modes of all port registers,
  - completely revamped how ports, timer pins, and interrupt pins all
    interoperate,
  - updated the VPU to render sprite priorities correctly,
  - added HDMA emulation,
  - fixed a bug in the ZCF instruction that was preventing the time from being
    set in the BIOS,
  - accidentally stumbled across a crude RAM patching hack so that subsequent
    boots don't keep dumping you into the BIOS setup (hopefully I can track down
    where the RAM variable is supposed to get set, and determine why higan's not
    setting it normally.)

+++ higan/msx/cpu/cpu.hpp
@@ -11,8 +11,8 @@ struct CPU : Z80, Z80::Bus, Thread {
   auto read(uint16 address) -> uint8 override;
   auto write(uint16 address, uint8 data) -> void override;
 
-  auto in(uint8 address) -> uint8 override;
-  auto out(uint8 address, uint8 data) -> void override;
+  auto in(uint16 address) -> uint8 override;
+  auto out(uint16 address, uint8 data) -> void override;
 
   //serialization.cpp
   auto serialize(serializer&) -> void;
